% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countElements.R
\name{countElements.NetworkDifferencesAspect}
\alias{countElements.NetworkDifferencesAspect}
\title{Number of elements in \link{NetworkDifferences} aspect}
\usage{
\method{countElements}{NetworkDifferencesAspect}(x)
}
\arguments{
\item{x}{an object of \link{NetworkDifferences} aspect class.}
}
\value{
integer; number of elements which is always 1.
}
\description{
This function returns the number of elements in a \link{NetworkDifferences} aspect.
}
\details{
The \code{\link[RCX:countElements]{RCX::countElements()}} function uses method dispatch, so the default methods already returns the correct number for the most aspect classes.
This way it is easier to extend the data model.
This function implements the function for the \link{NetworkDifferences} aspect.
}
\examples{
NULL
}
\seealso{
\code{\link[RCX:countElements]{RCX::countElements()}}, \code{\link[RCX:hasIds]{RCX::hasIds()}}, \code{\link[RCX:idProperty]{RCX::idProperty()}}, \code{\link[RCX:refersTo]{RCX::refersTo()}}, \code{\link[RCX:referredBy]{RCX::referredBy()}}, \code{\link[RCX:maxId]{RCX::maxId()}}
}
