% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNetworks.R
\name{.addAttributes}
\alias{.addAttributes}
\title{Helper function to add the nodes for the attributes}
\usage{
.addAttributes(rcx, offset, nodeName, belongsTo, shape, type)
}
\arguments{
\item{rcx}{\code{\link{RCX-object}} to which the nodes for the attributes are added}

\item{offset}{integer; id of the node which is needed for the nodeAttributes}

\item{nodeName}{character; name for the new node}

\item{belongsTo}{character; if the node belongs to the left, right or both networks}

\item{shape}{character; shape of the node that indicates the nodeType (nodeAttribute or edgeAttribute)}

\item{type}{character; if the attributes has different values in the two networks (noSharedValue or sharedValue)}
}
\value{
\code{\link{RCX-object}}
}
\description{
Helper function to add the nodes for the attributes
}
