% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNetworks.R
\name{.diffAddEdgeAttributes}
\alias{.diffAddEdgeAttributes}
\title{Helper-function to add nodes for the edgeAttributes to a \code{\link{RCX-object}}}
\usage{
.diffAddEdgeAttributes(
  rcx,
  edgeAttrDf = NULL,
  numberOfNodes = 0,
  startLayerValues = 0
)
}
\arguments{
\item{rcx}{\code{\link{RCX-object}}; the new created nodes are added to this \code{\link{RCX-object}}}

\item{numberOfNodes}{integer (optional, default value 0); parameter to know at which id the nodes for the edges start (needed for the function \emph{rcxNodeEdgeNetwork})}

\item{startLayerValues}{integer (optional, default value 0); if greater than 0, nodes for the values of the edgeAttributes are included}

\item{nodeAttrDf}{dataframe; contains the information for the added nodes for the names and values of edgeAttributes}
}
\value{
\code{\link{RCX-object}}
}
\description{
Add nodes for edgeAttributes to an existing \code{\link{RCX-object}} that visualizes the differences of two RCX-objects.
}
