% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNetworks.R
\name{.diffAddNodeAttributes}
\alias{.diffAddNodeAttributes}
\title{Helper-function to add nodes for the nodeAttributes to an RCX-object}
\usage{
.diffAddNodeAttributes(rcx, nodeAttrDf = NULL, startLayerValues = 0)
}
\arguments{
\item{rcx}{RCX-object; the new created nodes are added to this RCX-object}

\item{nodeAttrDf}{dataframe; contains the information for the new nodes}

\item{startLayerValues}{integer (optional, default value 0); if greater than 0, nodes for the values of the nodeAttributes are included}
}
\value{
RCX-object
}
\description{
Helper-function to add nodes for the nodeAttributes to an RCX-object
}
