% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNetworks.R
\name{.diffAddVisualization}
\alias{.diffAddVisualization}
\title{Helper-function to define the visualization for the \code{\link{RCX-object}}}
\usage{
.diffAddVisualization(rcx)
}
\arguments{
\item{rcx}{\code{\link{RCX-object}}; the layout is added to this \code{\link{RCX-object}}}
}
\value{
\code{\link{RCX-object}}
}
\description{
Helper-function to define the visualization for the \code{\link{RCX-object}}
}
\details{
Node shape in the visualization: Node name: round, node represent: triangle, edge: rectangle, name of nodeAttribute: hexagon,
value of nodeAttribute: parallelogram, name of edgeAttribute: rectangle with round corner, value of edgeAttribute: diamond
Color of the network-elements: gray: belongs to both networks, light blue: belongs only to the left network, orange: belongs only to the right network
}
