% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNetworks.R
\name{.diffGetMaxInLayers}
\alias{.diffGetMaxInLayers}
\title{Helper function for \emph{.diffQuadrantLayout}}
\usage{
.diffGetMaxInLayers(layer)
}
\description{
Helper function for \emph{.diffQuadrantLayout}
}
