% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNetworks.R
\name{.diffSortedLayout}
\alias{.diffSortedLayout}
\title{Helper-function to define to node positions for the \code{\link{RCX-object}}}
\usage{
.diffSortedLayout(
  rcx,
  dX = 70,
  dY = 70,
  startLayerBoth = 5,
  startLayerLeftRight = 10,
  startLayerAttributes = 0,
  startLayerValues = 0
)
}
\arguments{
\item{rcx}{\code{\link{RCX-object}}; the defined node positions are added to this \code{\link{RCX-object}}}

\item{dX}{integer (optional, default value 70); determines the width of the created circles}

\item{dY}{integer (optional, default value 70); determines the height of the created circles}

\item{startLayerBoth}{integer (optional, default value 5); determines the layer at which the circles for the nodes that
belong to both RCX-objects start}

\item{startLayerLeftRight}{integer (optional, default value 10); determines the layer at which the circles for the nodes that belong only
to one RCX-object start}

\item{startLayerAttributes}{integer (optional, default value 0); if greater than \emph{startLayerLeftRight}, it determines the layer at which
the circles for the names of the nodesAttributes start}

\item{startLayerValues}{integer (optional, default value 0); if greater than \emph{startLayerAttributes}, it determines the layer at which
the circles for the values of the nodesAttributes start}
}
\description{
Helper-function to define to node positions for the \code{\link{RCX-object}}
}
\details{
The layout of the \code{\link{RCX-object}} consists of concentric circles. The layers for the nodes/edges that exist in both network
are defined through the parameter \emph{startLayerBoth}. The layers for the nodes/edges that exist only in one network are defined through the parameter
\emph{startLayerLeftRight}. If \emph{startLayerAttributes} is greater than \emph{startLayerLeftRight}, the layers for the names of the node-/edgeAttributes
start at \emph{startLayerAttributes}. If \emph{startLayerValues} is greater than \emph{startLayerAttributes}, the layers for the values of the node-/edgeAttributes
start at \emph{startLayerValues}.
}
