% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{.jsonV}
\alias{.jsonV}
\title{Return data as a vector from a JSON list}
\usage{
.jsonV(data, acc, default = NA, returnAllDefault = TRUE)
}
\arguments{
\item{data}{json list}

\item{acc}{accession name}

\item{default}{default return value}

\item{returnAllDefault}{whether to return the vector if all values are the default value (or \code{NULL} instead)}
}
\value{
vector
}
\description{
Return data as a vector from a JSON list
}
\note{
Internal function only for convenience
}
\keyword{internal}
