% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNetworks.R
\name{exportDifferencesToNodeNetwork}
\alias{exportDifferencesToNodeNetwork}
\title{Node-centered \code{\link{RCX-object}} for the differences of two RCX-objects}
\usage{
exportDifferencesToNodeNetwork(
  aspect,
  includeNamesAndRepresents = FALSE,
  dX = 70,
  dY = 70,
  startLayerBoth = 5,
  startLayerLeftRight = 10,
  startLayerAttributes = 0,
  startLayerValues = 0
)
}
\arguments{
\item{aspect}{networkDifferences-aspect; information about the differences of two \code{\link{RCX-object}}}

\item{includeNamesAndRepresents}{logical (optional, default value \code{FALSE}); if set to \code{TRUE} nodes for node names and node represents are created}

\item{dX}{integer (optional, default value 70); determines the width of the created circles}

\item{dY}{integer (optional, default value 70); determines the height of the created circles}

\item{startLayerBoth}{integer (optional, default value 5); determines the layer at which the circles for the nodes that
belong to both RCX-objects start}

\item{startLayerLeftRight}{integer (optional, default value 10); determines the layer at which the circles for the nodes that belong only
to one RCX-object start}

\item{startLayerAttributes}{integer (optional, default value 0); if greater than \emph{startLayerLeftRight}, it determines the layer at which
the circles for the names of the nodesAttributes start}

\item{startLayerValues}{integer (optional, default value 0); if greater than \emph{startLayerAttributes}, it determines the layer at which
the circle for the values of the nodesAttribute start}
}
\value{
\code{\link{RCX-object}}
}
\description{
Creation of a node-centered \code{\link{RCX-object}} with the information of a \emph{networkDifferences}-aspect to visualize
the differences of two RCX-objects.
}
\details{
A \code{\link{RCX-object}} is created with nodes for the node names / node represents (depending on \emph{includeNamesAndRepresents}),
nodes for the names of the nodeAttributes (if \emph{startLayerAttributes} greater than \emph{startLayerLeftRight}), nodes for the values of the nodeAttributes
(if \emph{startLayerValues} is greater than \emph{startLayerAttributes}). The following inequation must be satisfied: 0 < \emph{startLayerBoth} < \emph{startLayerLeftRight} < \emph{startLayerAttributes} < \emph{startLayerValues}.

The networkAttributes are ignored as they are not considered to very important.

Node shape in the visualization: Node name: round, node represent: triangle, name of nodeAttribute: hexagon, value of nodeAttribute: parallelogram.

Color of the network-elements: gray: belongs to both networks, light blue: belongs only to the left network, orange: belongs only to the right network.

If there are edges between two nodes in the left or right RCX-objects, colored edges are defined in the created RCX-object, too, and are labeled with the interaction of the edge.
If \emph{includeNamesAndRepresents} is \code{TRUE}, edges between the nodes for the node and node represent are defined.
If nodes for the names of the nodeAttributes are included, edges between these nodes and the node they belong to are defined.
If nodes for the values of the nodeAttributes are included, edges between these nodes and the nodes for the names of the nodeAttributes are defined.
}
