% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.NetworkDifferencesAspect}
\title{Printing the NetworkDifferences aspect}
\usage{
\method{print}{NetworkDifferencesAspect}(x)
}
\arguments{
\item{x}{NetworkDifferences aspect}

\item{...}{further arguments passed to or from other methods. See \code{\link[base:print]{base::print()}}}
}
\value{
prints the NetworkDifferences aspect and returns it invisibly (\href{x}{invisible})
}
\description{
These functions attempt to print the aspect of an RCX object in a more readable form.
}
