% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{rcxToJson.NetworkDifferencesAspect}
\alias{rcxToJson.NetworkDifferencesAspect}
\title{Convert \link{NetworkDifferences} aspects to JSON}
\usage{
\method{rcxToJson}{NetworkDifferencesAspect}(aspect, verbose = FALSE)
}
\arguments{
\item{aspect}{\link{NetworkDifferences} aspect}

\item{verbose}{logical; whether to print what is happening}

\item{...}{additional parameters, that might needed for extending}
}
\value{
character; JSON of an aspect
}
\description{
Functions for converting the different aspects to JSON following the CX data structure definition
(see NDEx documentation: \url{https://home.ndexbio.org/data-model/}).
}
\details{
For converting \link[RCX:createRCX]{RCX objects} to JSON, each aspect is processed by a generic function for its
aspect class. Those functions return a character only containing the JSON of this aspect, which is then
combined by \link[RCX:toCX]{RCX::toCX} to be a valid CX data structure.

To support the conversion for non-standard or own-defined aspects, generic functions for those aspect classes
have to be implemented.
}
\seealso{
\link[RCX:toCX]{RCX::toCX}, \link[RCX:writeCX]{RCX::writeCX}, \link[RCX:jsonToRCX]{RCX::jsonToRCX}, \link[RCX:readCX]{RCX::readCX}
}
