% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNetworks.R
\name{.diffAddNodes}
\alias{.diffAddNodes}
\title{Helper-function to add nodes to a \code{\link{RCX-object}}
\usage{
.diffAddNodes(
  rcx = NULL,
  nodeDf = NULL,
  matchByName = TRUE,
  includeNamesAndRepresents = FALSE
)
}
\arguments{
\item{rcx}{\code{\link{RCX-object}}; the new nodes are added to this \code{\link{RCX-object}}}

\item{nodeDf}{dataframe; contains the information for the new nodes}

\item{matchByName}{logical (optional, default value \code{TRUE}); if set to \code{TRUE}, two nodes are equal when their names are equal,
if set to \code{FALSE}, two nodes are equal when their represents are equal}

\item{includeNamesAndRepresents}{logical (optional, default value \code{FALSE}); if set to \code{TRUE}, nodes for node names and
node represents are created}
}
\value{
\code{\link{RCX-object}
}
\description{
Add nodes for nodes names / represents to an existing \code{\link{RCX-object}} that visualizes the differences of two RCX-objects.
}
