% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNetworks.R
\name{exportDifferencesToEdgeNetwork}
\alias{exportDifferencesToEdgeNetwork}
\title{Edge-centered \code{\link{RCX-object}} for the differences of two RCX-objects}
\usage{
exportDifferencesToEdgeNetwork(
  aspect = NULL,
  dX = 70,
  dY = 70,
  startLayerBoth = 5,
  startLayerLeftRight = 10,
  startLayerAttributes = 0,
  startLayerValues = 0
)
}
\arguments{
\item{aspect}{networkDifferences-aspect; information about the differences of two \code{\link{RCX-object}}}

\item{dX}{integer (optional, default value 70); determines the width of the created circles}

\item{dY}{integer (optional, default value 70); determines the height of the created circles}

\item{startLayerBoth}{integer (optional, default value 5); determines the layer at which the circles for the edges that belong to both RCX-object start}

\item{startLayerLeftRight}{integer (optional, default value 10); determines the layer at which the circles for the nodes that belong only to one RCX-object start}

\item{startLayerAttributes}{integer (optional, default value 0); if greater than *startLayerLeftRight*, it determines the layer at which 
the circles for the names of the nodesAttributes start}

\item{startLayerValues}{integer (optional, default value 0); if greater than *startLayerAttributes*, it determines the layer at which 
the circle for the values of the nodesAttribute start}
}
\value{
\code{\link{RCX-object}}
}
\description{
Creation of an edge-centered RCX-object with the information of a *networkDifferences*-aspect to visualize 
the differences of two RCX-objects.
rcxNodeNetwork
}
\details{
A \code{\link{RCX-object}} is created with nodes for the edge interaction, nodes for the names of the edgeAttributes 
(if *startLayerAttributes* greater than *startLayerLeftRight*), nodes for the values of the edgeAttributes (if *startLayerValues* is greater than *startLayerAttributes*). 
The following inequation must be satisfied: 0 < *startLayerBoth* < *startLayerLeftRight* < *startLayerAttributes* < *startLayerValues*.
  
The networkAttributes are ignored as they are not considered to very important.

Node shape in the visualization: Edge: rectangle, name of edgeAttribute: rectangle with round corner, value of edgeAttribute: diamond.

Color of the network-elements: gray: belongs to both networks, light blue: belongs only to the left network, orange: belongs only to the right network.

If in the left or right RCX-object two edges have a node in common, colored edges are defined in the created RCX-object, 
and the edges are labeled with the node names / node represent (depending on *matchByName*).
If nodes for the names of the edgeAttributes are included, edges between these nodes and the nodes for the edge they belong to are defined.
If nodes for the values of the edgeAttributes are included, edges between these nodes and the nodes for the names of the edgeAttributes are defined.
}
