% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.NetworkDifferencesAspect}
\title{Printing the \link{NetworkDifferences} aspect}
\usage{
\method{print}{NetworkDifferencesAspect}(x)
}
\arguments{
\item{x}{\link{NetworkDifferences} aspect}
}
\value{
prints the \link{NetworkDifferences} aspect and returns it invisibly (\href{x}{invisible})
}
\description{
These functions attempt to print the aspect of an RCX object in a more readable form.
}
