% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updateNetworkDifferences}
\alias{updateNetworkDifferences}
\alias{updateNetworkDifferences.NetworkDifferencesAspect}
\alias{updateNetworkDifferences.RCX}
\title{Update RCX-object with NetworkDifferences aspect}
\usage{
updateNetworkDifferences(x, aspect)

\method{updateNetworkDifferences}{NetworkDifferencesAspect}(x, aspect, replace = TRUE, ...)

\method{updateNetworkDifferences}{RCX}(x, networkDifferences, ...)
}
\arguments{
\item{x}{\link[RCX:createRCX]{RCX object}; (to which the new \link{networkDifferences} will be added)}

\item{...}{additional parameters}

\item{networkDifferences}{\link{NetworkDifferences} aspect}
}
\value{
updated \link[RCX:createRCX]{RCX object}
}
\description{
This functions updates an \link[RCX:createRCX]{RCX object} with a \link{NetworkDifferences} aspect
}
