% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updateNetworkDifferences}
\alias{updateNetworkDifferences}
\title{Update networkDifferences}
\usage{
updateNetworkDifferences(x, networkDifferences, ...)
}
\arguments{
\item{x}{\link[RCX:createRCX]{RCX object} or \link{networkDifferences} object; (to which the new \link{networkDifferences} will be added)}

\item{networkDifferences}{\link{networkDifferences} object; (the \link{networkDifferences} aspect, that will be added)}

\item{...}{additional parameters}
}
\value{
updated \link[RCX:createRCX]{RCX object} or \link{MetaRelSubNetVis} object
}
\description{
This functions updates an \link{networkDifferences} aspect with an other \link{networkDifferences} aspect or within an \link[RCX:createRCX]{RCX object}.
}
\seealso{
\link[RCX:updateNodeAttributes]{RCX::updateNodeAttributes}, \link[RCX:updateEdgeAttributes]{RCX::updateEdgeAttributes}, \link[RCX:updateNetworkAttributes]{RCX::updateNetworkAttributes}
}
